reset db uses initstats (mostly)

+++ src/functions/database/database.seeder.ts
@@ -7,13 +7,18 @@ import roundModel from "./models/round.model";
 import buildingModel from "./models/building.model";
 import alertModel from "./models/alert.model";
 import { getAllTiles } from "./database.server";
-import { Corporation, Tile } from "@/types";
+import { Corporation, Tile, RESET_DB_ACTION } from "@/types";
+import initialcorpsModel from "./models/initialstats/initialcorps.model";
+import initialbuildingModel from "./models/initialstats/initialbuilding.model";
+import mongoose from "mongoose";
 
 //delete all models
-export const deleteAllModels = async () => {
+export const deleteAllModels = async (action: RESET_DB_ACTION) => {
   await roundModel.deleteMany();
-  await tileModel.deleteMany();
-  await alertModel.deleteMany();
+  if (action === RESET_DB_ACTION.SEED) {
+    await tileModel.deleteMany();
+    await alertModel.deleteMany();
+  }
   await buildingModel.deleteMany();
   await corporationModel.deleteMany();
 }
@@ -45,7 +50,7 @@ export async function createAllCorporations() {
 }
 
 export const createAllTiles = async () => {
-  const tileFixtures : Tile[] = getFixtureTiles();
+  const tileFixtures: Tile[] = getFixtureTiles();
   //tiles
   await tileModel.deleteMany();
   const newTiles = await tileModel.create(tileFixtures);
@@ -54,7 +59,7 @@ export const createAllTiles = async () => {
 
 export const seedDB = async () => {
 
-  await deleteAllModels();
+  await deleteAllModels(RESET_DB_ACTION.SEED);
 
   //round reset
   await roundModel.create(firstRound);
@@ -70,3 +75,40 @@ export const seedDB = async () => {
 
   console.log("================== DATABASE SEEDED ==================");
 };
+
+export const resetDB = async () => {
+  await deleteAllModels(RESET_DB_ACTION.GAME_RESET);
+
+  //round reset
+  const initBuildings = await initialbuildingModel.find();
+  for(const building of initBuildings) {
+    building._id = new mongoose.Types.ObjectId();
+    building.isNew = true;
+    await buildingModel.create(building);
+  }
+
+  const initCorporations = await initialcorpsModel.find();
+  for(const corporation of initCorporations) {
+    corporation._id = new mongoose.Types.ObjectId();
+    corporation.isNew = true;
+    await corporationModel.create(corporation);
+  }
+
+  // resources?
+
+  let round = await roundModel.findOne();
+  if(!round) round = await roundModel.create({});
+  round = await roundModel.findOne();
+  round.number = 0;
+  round.playing = false;
+  round.darkHour = true;
+  await round.save();
+
+  const tiles = await tileModel.find();
+  for(const tile of tiles) {
+    tile.colonizedBy = null;
+    tile.buildings = [];
+    tile.destroyed = false;
+    await tile.save();
+  }
+}
